requirements.txt
black==22.8.0
click==8.1.3
crytic-compile==0.2.3
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
prettytable==3.4.1
pysha3==1.0.2
slither-analyzer==0.8.3
solc-select==0.2.1
tomli==2.0.1
typing-extensions==4.3.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
